게시물 작성시 작성자 정보 저장, 수정/삭제 권한체크

+++ src/main/java/com/KoreaIT/java/jam/ArticleDoDeleteServlet.java
@@ -4,12 +4,14 @@ import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
+import java.util.Map;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.KoreaIT.java.jam.config.Config;
 import com.KoreaIT.java.jam.util.DBUtil;
@@ -20,6 +22,13 @@ public class ArticleDoDeleteServlet extends HttpServlet {
        
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		response.setContentType("text/html;charset=UTF-8");
+		
+		HttpSession session = request.getSession();	
+		
+		if(session.getAttribute("loginedMemberId")==null) {
+			response.getWriter().append(String.format("<script>alert('로그인 후 이용해주세요.'); location.replace('../article/list')</script>"));
+			return;
+		}
 		
 		// DB 연결
 		Connection conn = null;
@@ -45,9 +54,20 @@ public class ArticleDoDeleteServlet extends HttpServlet {
 			
 			int id = Integer.parseInt(inputedid);
 			
-			SecSql sql = new SecSql();
+			SecSql sql = SecSql.from("SELECT *");
+			sql.append("FROM article");
+			sql.append("WHERE id = ? ;", id);
+
+			Map<String, Object> articleRow = DBUtil.selectRow(conn, sql);
+
+			if(session.getAttribute("loginedMemberId")!=articleRow.get("memberId")) {
+				response.getWriter().append(String.format("<script>alert('이 게시글에 권한이 없습니다.'); location.replace('../article/list')</script>"));
+				return;
+			}
+			
+			sql = new SecSql();
 			sql.append("DELETE FROM article");
-			sql.append("WHERE id = ?", id);
+			sql.append("WHERE id = ?;", id);
 			
 			DBUtil.delete(conn, sql);
